Tournament works in batch mode, but GUI still needs work.
Changed the include directory to lib for collecting external libraries.

Added a TclRobots version of the puts command where you can set the
level of the message for visibility control with startup flags. The
common puts is renamed tcl_puts, which must be used in cases where
channels other than stdout is needed.